NOTE:  This is on branch pking.templated_beamline.
Merged trunk changes from revision range 1202:1481 into branch 
pking.templated_beamline (in two chunks: 1202:1472 and 1472:1481 ).
There were conflicts found and resolved in the first chunk; no conflicts
in the second chunk.


git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/branches/pking.templated_beamline@1482 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Analysis/include/StringManip.h
@@ -19,7 +19,7 @@ std::string get_line_seg(const std::string& stmp, const std::string& delim1, con
 
 std::string get_line_seg(const std::string& stmp, const std::string& delim);
 
-std::string get_next_seg(const std::string& inputstring, unsigned int& pos);
+std::string get_next_seg(const std::string& inputstring, size_t& pos);
 
 void split_string(const std::string& inputstring, const std::string& delim, std::vector<std::string>& varlist);
